requirements.txt
flask
gensim
numpy
python-Levenshtein